oracle/pyproject.toml
[tool.poetry]
name = "oracle"
version = "0.1.0"
description = ""
authors = ["Taehoon Kim, <taehoon.kim@uzh.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^6.19.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
numpy = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
